pyproject.toml
[project]
name = "ec-issuer"
version = "0.3.0"
description = "Educredentials Issuer"
authors = [{ name = "Bèr Kessels", email = "ber@berk.es" }]
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "jwcrypto>=1.5.7",
    "msgspec>=0.20.0",
    "prometheus-flask-exporter>=0.23.2",
    "psycopg2-binary>=2.9.12",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "basedpyright>=1.38.4",
    "jsonpath-ng>=1.8.0",
    "jsonschema>=4.26.0",
    "pdoc>=16.0.0",
    "pyjwt[crypto]>=2.12.1",
    "pytest>=9.0.3",
    "pytest-cov>=4.0.0",
    "ruff>=0.15.4",
    "types-requests>=2.32.4.20260107",
]

[tool.ruff.lint]
select = ["E", "F"]
exclude = ["Justfile"]

[tool.basedpyright]
typeCheckingMode = "recommended"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
markers = [
    "e2e: end-to-end tests",
    "integration: integration tests",
    "unit: unit tests",
]
